core/src/test/java/shop/core/order/OrderServiceImplTest.java
package shop.core.order;

import org.junit.jupiter.api.Test;
import shop.core.discount.FixDiscountPolicy;
import shop.core.member.Grade;
import shop.core.member.Member;
import shop.core.member.MemberRepository;
import shop.core.member.MemoryMemberRepository;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

class OrderServiceImplTest {

    @Test
    void createOrder() {
        MemoryMemberRepository memoryMemberRepository = new MemoryMemberRepository();
        memoryMemberRepository.save(new Member(1L, "name", Grade.VIP));

        OrderServiceImpl orderService = new OrderServiceImpl(memoryMemberRepository, new FixDiscountPolicy());
        Order order = orderService.createOrder(1L, "itemA", 10000);
        assertThat(order.getDiscountPrice()).isEqualTo(1000);
    }

}